+++ frontend/pages/sites/_siteId.vue
@@ -165,6 +165,10 @@ export default class SitePage extends Vue {
         name: 'Contacts'
       },
       {
+        to: '/sites/' + this.siteId + '/locations',
+        name: 'Locations'
+      },
+      {
         to: '/sites/' + this.siteId + '/attachments',
         name: 'Attachments'
       }